move things out of jwt.js
