+++ packages/api/src/utils/jwt.js
@@ -1,5 +1,3 @@
-import { base58btc } from 'multiformats/bases/base58'
-
 /** @type {Record<string, any>} */
 const algorithms = {
   HS256: {
@@ -8,17 +6,10 @@ const algorithms = {
       name: 'SHA-256',
     },
   },
-
-  // NODE-ED25519 exists only in the CloudFlare workers runtime API
-  // see: https://developers.cloudflare.com/workers/runtime-apis/web-crypto#footnote%201
-  'NODE-ED25519': {
-    name: 'NODE-ED25519',
-    namedCurve: 'NODE-ED25519',
-  },
 }
 
 // Adapted from https://chromium.googlesource.com/chromium/blink/+/master/LayoutTests/crypto/subtle/hmac/sign-verify.html
-var Base64URL = {
+export var Base64URL = {
   stringify: function (/** @type {Uint8Array} */ a) {
     // @ts-ignore
     var base64string = btoa(String.fromCharCode.apply(0, a))
@@ -48,7 +39,7 @@ function isString(s) {
 /**
  * @param {string | number | boolean} str
  */
-function utf8ToUint8Array(str) {
+export function utf8ToUint8Array(str) {
   str = btoa(unescape(encodeURIComponent(str)))
   return Base64URL.parse(str)
 }
@@ -107,119 +98,6 @@ export async function verifyJWT(token, secret, alg = 'HS256') {
   )
 
   return Base64URL.stringify(new Uint8Array(signedHeaderPayload)) === signature
-}
-
-/**
- * @param {string} token - an encoded JWT token from an x-web3auth header.
- * @returns {Promise<boolean>} - true if the token contains a valid solana public key and a valid signature from the key.
- */
-export async function verifyMetaplexJWT(token) {
-  if (!isString(token)) {
-    throw new Error('token must be a string')
-  }
-  var tokenParts = token.split('.')
-
-  if (tokenParts.length !== 3) {
-    return false
-  }
-
-  const header = parseJWTHeader(token)
-  const payload = parseJWT(token)
-
-  if (header.alg !== 'EdDSA') {
-    throw new Error('invalid algorithm for metaplex token')
-  }
-
-  if (header.typ !== 'JWT') {
-    throw new Error('invalid token type')
-  }
-
-  if (!payload.iss) {
-    return false
-  }
-
-  const pubkey = await keyFromDID(payload.iss)
-  const sig = Base64URL.parse(tokenParts[2])
-  const headerPayload = utf8ToUint8Array(tokenParts[0] + '.' + tokenParts[1])
-
-  if (!crypto || !crypto.subtle) {
-    return verifyEd25519SignatureWithJSCrypto(headerPayload, sig, pubkey)
-  }
-
-  try {
-    return await verifyEd25519SignatureWithNativeCrypto(
-      headerPayload,
-      sig,
-      pubkey
-    )
-  } catch (e) {
-    if (e instanceof Error && e.name === 'NotSupportedError') {
-      return await verifyEd25519SignatureWithJSCrypto(
-        headerPayload,
-        sig,
-        pubkey
-      )
-    }
-    throw e
-  }
-}
-
-/**
- *
- * @param {Uint8Array} message
- * @param {Uint8Array} sig
- * @param {Uint8Array} pubkey
- * @returns {Promise<boolean>}
- */
-async function verifyEd25519SignatureWithNativeCrypto(message, sig, pubkey) {
-  console.log()
-  const key = await crypto.subtle.importKey(
-    'raw',
-    pubkey,
-    algorithms['NODE-ED25519'],
-    false,
-    ['verify']
-  )
-
-  return crypto.subtle.verify(algorithms['NODE-ED25519'], key, sig, message)
-}
-
-/**
- *
- * @param {Uint8Array} message
- * @param {Uint8Array} sig
- * @param {Uint8Array} pubkey
- * @returns {Promise<boolean>}
- */
-async function verifyEd25519SignatureWithJSCrypto(message, sig, pubkey) {
-  console.warn(
-    'using tweetnacl for ed25519 - you should not see this message when running in the CloudFlare worker runtime'
-  )
-  const { default: nacl } = await import('tweetnacl')
-  return nacl.sign.detached.verify(message, sig, pubkey)
-}
-
-/**
- * @param {string} did - a "did:key" formatted DID string
- * @returns {Promise<Uint8Array>} - the decoded public key
- * @throws if DID is invalid or does not contain a valid Ed25519 public key
- */
-async function keyFromDID(did) {
-  const prefix = 'did:key:'
-  if (!did.startsWith(prefix)) {
-    throw new Error('invalid DID')
-  }
-
-  const keyStr = did.slice(prefix.length)
-  const bytes = base58btc.decode(keyStr)
-
-  // check multicodec == ed25519-pub (0xed encoded as a varint)
-  if (bytes[0] !== 0xed || bytes[1] !== 0x01) {
-    throw new Error(
-      'invalid key multicodec. only ed25519-pub keys are supported'
-    )
-  }
-  return bytes.slice(2)
 }
 
 /**
@@ -329,6 +207,6 @@ export function parseJWT(token) {
  * @param {string} token
  * @returns {JWTHeader}
  */
-function parseJWTHeader(token) {
+export function parseJWTHeader(token) {
   return JSON.parse(decodeJWTHeader(token))
 }
